Store block delay on chain contract (#73)

+++ consensus-client-it/src/test/scala/units/C2ENativeTokenTransfersViaDepositsTestSuite.scala
@@ -57,7 +57,7 @@ class C2ENativeTokenTransfersViaDepositsTestSuite extends BaseDockerTestSuite {
     waves1.api.broadcastAndWait(
       TxHelpers.dataEntry(
         chainContractAccount,
-        IntegerDataEntry("strictC2ETransfersActivationEpoch", activationEpoch)
+        IntegerDataEntry("strictC2ETransfersActivationEpoch", activationEpoch.toInt)
       )
     )
